requirements.txt
ccxt>=4.4.0
pandas>=2.2.0
requests>=2.32.0
websocket-client>=1.8.0
python-dotenv>=1.0.1
